+++ changelog.md
@@ -1,2 +1,8 @@
 ## [1.0.0 (January 7th, 2024)](#1-0-0)
 - Initial Release
+
+## [1.0.1 (January 8th, 2024)](#1-0-1)
+- Fixed Compatability issues between version 1.4.1 of Cobblemon
+
+## [1.0.2 (January 8th, 2024)](#1-0-2)
+- Fixed Server functionality